requirements.txt
# sudo apt-get install mysql-server libmysqlclient-dev libpq-dev python-dev python3-dev build-essential 
# dvc init - to track dataset 
# dvc remote add -d storage gdrive://1zOuuO3tVDjAwvJV9Bowk5-hdNFiL93Y1 - set remote storage gdrive
# dvc add /home/asdf/prj/MLFlow/data/LargeMovieReviewDataset.csv - add dataset to the dvc tracking
# dvc push - create link with a dvc remote sta folder with data in gdrive girectory
# rm -f /home/asdf/prj/MLFlow/data/LargeMovieReviewDataset.csv
# rm -rf .dvc/cache
# dvc pull - to pull dataset from gdrive 
# sudo apt install openjdk-11-jre-headless
# sudo apt install docker-compose

wheel == 0.37.1
sklearn-pandas == 2.2.0
scripts == 2.0
unidecode == 1.3.4
word2number == 1.1
contractions == 0.1.72
pyspellchecker==0.5.6
pattern == 3.6
beautifulsoup4 == 4.11.1
nltk == 3.7
spacy == 2.3.5
unidecode == 1.3.4
matplotlib == 3.5.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
hyperopt == 0.2.7
seaborn == 0.11.2
mlflow == 1.27.0
gevent == 21.12.0
dvc[gdrive] == 2.13.0
pyspark
py4j==0.10.9.5
dvc-render==0.0.6
pyOpenSSL==20.0.1
cryptography==36.0.2

